+++ tests/test_device_scanner_errors.py
@@ -1,139 +1,73 @@
-"""Error-path and logging device scanner tests."""
-
-import logging
-from unittest.mock import AsyncMock, MagicMock, patch
+from unittest.mock import AsyncMock, patch
 
 import pytest
-from custom_components.thessla_green_modbus.const import CONNECTION_MODE_TCP, SENSOR_UNAVAILABLE
 from custom_components.thessla_green_modbus.scanner.core import ThesslaGreenDeviceScanner
 
-pytestmark = pytest.mark.asyncio
+
+async def _make_scanner(**kwargs):
+    return await ThesslaGreenDeviceScanner.create("192.168.1.1", 502, 1, **kwargs)
 
 
-async def test_log_invalid_value_debug_when_not_verbose(caplog):
-    """Invalid values log at DEBUG level when not verbose."""
-    scanner = ThesslaGreenDeviceScanner("host", 502)
+@pytest.mark.asyncio
+async def test_scan_device_scan_returns_non_dict_raises():
+    scanner = await _make_scanner()
 
-    caplog.set_level(logging.DEBUG)
-    scanner._log_invalid_value("test_register", 1)
+    async def fake_scan(self_arg):
+        return "not_a_dict"
 
-    assert caplog.records[0].levelno == logging.DEBUG
-    assert "Invalid value for test_register: raw=1 decoded=1" in caplog.text
+    with patch.object(ThesslaGreenDeviceScanner, "scan", fake_scan):
+        with patch.object(scanner, "close", AsyncMock()):
+            with pytest.raises(TypeError, match="scan\\(\\) must return a dict"):
+                await scanner.scan_device()
 
-    caplog.clear()
-    scanner._log_invalid_value("test_register", 1)
 
-    assert not caplog.records
+@pytest.mark.asyncio
+async def test_scan_device_main_scan_non_dict_raises():
+    scanner = await _make_scanner()
 
-async def test_log_invalid_value_info_then_debug_when_verbose(caplog):
-    """First invalid value logs INFO when verbose, then DEBUG."""
-    scanner = ThesslaGreenDeviceScanner("host", 502, verbose_invalid_values=True)
+    async def fake_scan_non_dict(self_arg):
+        return "not_a_dict"
 
-    caplog.set_level(logging.DEBUG)
-    scanner._log_invalid_value("test_register", 1)
+    mock_transport = AsyncMock()
+    mock_transport.client = AsyncMock()
 
-    assert caplog.records[0].levelno == logging.INFO
-    assert "raw=1" in caplog.text
+    with patch.object(ThesslaGreenDeviceScanner, "scan", fake_scan_non_dict):
+        with patch.object(scanner, "_build_tcp_transport", return_value=mock_transport):
+            with patch.object(scanner, "close", AsyncMock()):
+                with pytest.raises(TypeError, match="scan\\(\\) must return a dict"):
+                    await scanner.scan_device()
 
-    caplog.clear()
-    scanner._log_invalid_value("test_register", 1)
 
-    assert caplog.records[0].levelno == logging.DEBUG
-    assert "raw=1" in caplog.text
+@pytest.mark.asyncio
+async def test_scan_device_legacy_returns_dict():
+    scanner = await _make_scanner()
 
-async def test_log_invalid_value_raw_and_formatted(caplog):
-    """Log includes both raw hex and decoded representation."""
-    scanner = ThesslaGreenDeviceScanner("host", 502)
+    async def fake_scan_returns_dict():
+        return {"custom_key": "custom_value"}
 
-    caplog.set_level(logging.DEBUG)
-    scanner._log_invalid_value("schedule_time", 5632)
+    scanner.scan = fake_scan_returns_dict  # type: ignore[method-assign]
 
-    assert "raw=5632" in caplog.text
-    assert "decoded=16:00" in caplog.text
+    with patch.object(scanner, "close", AsyncMock()):
+        result = await scanner.scan_device()
 
-async def test_log_invalid_value_invalid_time(caplog):
-    """Logs include formatted string for invalid time values."""
-    scanner = ThesslaGreenDeviceScanner("host", 502)
+    assert result == {"custom_key": "custom_value"}
 
-    caplog.set_level(logging.DEBUG)
-    scanner._log_invalid_value("schedule_time", 9216)
 
-    assert "raw=9216" in caplog.text
-    assert "decoded=9216 (invalid)" in caplog.text
+@pytest.mark.asyncio
+async def test_scan_device_importlib_fails():
+    scanner = await _make_scanner()
+    mock_transport = AsyncMock()
+    mock_transport.client = AsyncMock()
 
-async def test_missing_register_logged_once(caplog):
-    """Each missing register should trigger only one read and log entry."""
-    empty_regs = {4: {}, 3: {}, 1: {}, 2: {}}
-    with patch.object(
-        ThesslaGreenDeviceScanner,
-        "_load_registers",
-        AsyncMock(return_value=(empty_regs, {})),
+    with (
+        patch.object(scanner, "_build_tcp_transport", return_value=mock_transport),
+        patch.object(scanner, "_read_input_block", AsyncMock(return_value=[])),
+        patch.object(scanner, "_read_holding_block", AsyncMock(return_value=[])),
+        patch.object(scanner, "_read_input", AsyncMock(return_value=None)),
+        patch.object(scanner, "_read_holding", AsyncMock(return_value=None)),
+        patch.object(scanner, "_read_coil", AsyncMock(return_value=None)),
+        patch.object(scanner, "_read_discrete", AsyncMock(return_value=None)),
     ):
-        scanner = await ThesslaGreenDeviceScanner.create("192.168.1.1", 502, 10)
+        result = await scanner.scan_device()
 
-    call_log: list[tuple[int, int]] = []
-
-    async def fake_read_input(client, address, count, **kwargs):
-        call_log.append((address, count))
-        if address == 0 and count == 5:
-            return [4, 85, 0, 0, 0]
-        if address == 24 and count == 6:
-            return [0] * 6
-        if address == 1 and count == 2:
-            return None
-        if address == 1 and count == 1:
-            return [1]
-        if address == 2 and count == 1:
-            return None
-        return [0] * count
-
-    scanner._input_register_map = {"reg_ok": 1, "reg_missing": 2}
-    scanner._holding_register_map = {}
-    scanner._coil_register_map = {}
-    scanner._discrete_input_register_map = {}
-    scanner._known_missing_registers = {
-        "input_registers": set(),
-        "holding_registers": set(),
-        "coil_registers": set(),
-        "discrete_inputs": set(),
-    }
-    scanner._update_known_missing_addresses()
-
-    with patch("pymodbus.client.AsyncModbusTcpClient") as mock_client_class:
-        mock_client = AsyncMock()
-        mock_client.connect.return_value = True
-        mock_client.read_input_registers = AsyncMock(
-            return_value=MagicMock(isError=lambda: False, registers=[4, 85, 0, 0, 0])
-        )
-        mock_client_class.return_value = mock_client
-
-        with (
-            patch.object(scanner, "_read_input", AsyncMock(side_effect=fake_read_input)),
-            patch.object(scanner, "_read_holding", AsyncMock(return_value=[0])),
-            patch.object(scanner, "_read_coil", AsyncMock(return_value=[False])),
-            patch.object(scanner, "_read_discrete", AsyncMock(return_value=[False])),
-            patch.object(scanner, "_is_valid_register_value", return_value=True),
-        ):
-            caplog.set_level(logging.DEBUG)
-            scanner.connection_mode = CONNECTION_MODE_TCP
-            result = await scanner.scan_device()
-
-    # Block read + single read for each register
-    assert call_log.count((1, 2)) == 1
-    assert call_log.count((1, 1)) == 1
-    assert call_log.count((2, 1)) == 1
-
-    # Missing register logged only once
-    assert caplog.text.count("Failed to read input_registers register 2") == 1
-
-    # Only valid register is reported as available
-    assert "reg_ok" in result["available_registers"]["input_registers"]
-    assert "reg_missing" not in result["available_registers"]["input_registers"]
-
-async def test_temperature_unavailable_no_warning(caplog):
-    """SENSOR_UNAVAILABLE should not log a warning — register exists, sensor just not connected."""
-    scanner = await ThesslaGreenDeviceScanner.create("192.168.1.100", 502, 10)
-
-    caplog.set_level(logging.WARNING)
-    assert scanner._is_valid_register_value("outside_temperature", SENSOR_UNAVAILABLE) is True
-    assert "outside_temperature" not in caplog.text
+    assert "available_registers" in result
